stt-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3